refactor(app): Simplify app.go with less redundant imports using struct embedding.
